Rename MemberNotfoundException to SourceNotFoundException, Change the package of SourceNotFoundException and RequiredParamNotFoundException member to commonException
